+++ bmds/bmds3/constants.py
@@ -1,11 +1,8 @@
 from enum import Enum, IntEnum
-from itertools import chain
-from typing import List, Optional, Tuple
+from typing import Tuple
 
-import numpy as np
 from pydantic import BaseModel
 
-from ..utils import pretty_table
 
 BMDS_BLANK_VALUE = -9999
 CDF_TABLE_SIZE = 99
@@ -162,24 +159,6 @@ class PriorType(IntEnum):
     Lognormal = 2
 
 
-class Prior(BaseModel):
-    name: str
-    type: PriorType
-    initial_value: float
-    stdev: float
-    min_value: float
-    max_value: float
-
-    def tbl_str_hdr(self) -> str:
-        return "| param | type       |    initial |      stdev |        min |        max |"
-
-    def tbl_str(self) -> str:
-        return f"| {self.name:5} | {self.type.name:10} | {self.initial_value:10.3g} | {self.stdev:10.3g} | {self.min_value:10.3g} | {self.max_value:10.3g} |"
-
-    def numeric_list(self) -> List[float]:
-        return list(self.dict(exclude={"name"}).values())
-
-
 class PriorClass(IntEnum):
     frequentist_unrestricted = 0
     frequentist_restricted = 1
@@ -196,73 +175,3 @@ _pc_name_mapping = {
     PriorClass.frequentist_restricted: "Frequentist restricted",
     PriorClass.bayesian: "Bayesian",
 }
-
-
-class ModelPriors(BaseModel):
-    prior_class: PriorClass  # if this is a predefined model class
-    priors: List[Prior]  # priors for main model
-    variance_priors: Optional[List[Prior]]  # priors for variance model (continuous-only)
-
-    def __str__(self):
-        ps = [self.priors[0].tbl_str_hdr()]
-        ps.extend([p.tbl_str() for p in self.priors])
-        p = "\n".join(ps)
-        if self.variance_priors is not None:
-            vps = "\n".join([p.tbl_str() for p in self.variance_priors])
-            p += f"""\n{vps}"""
-        p += "\n"
-        return p
-
-    def tbl(self) -> str:
-        headers = "name|type|initial_value|stdev|min_value|max_value".split("|")
-        rows = [
-            (p.name, p.type.name, p.initial_value, p.stdev, p.min_value, p.max_value)
-            for p in chain(self.priors, self.variance_priors or ())
-        ]
-        return pretty_table(rows, headers)
-
-    def get_prior(self, name: str) -> Prior:
-        """Search all priors and return the match by name.
-
-        Args:
-            name (str): prior name
-
-        Raises:
-            ValueError: if no value is found
-        """
-        for p in self.priors:
-            if p.name == name:
-                return p
-        if self.variance_priors:
-            for p in self.variance_priors:
-                if p.name == name:
-                    return p
-        raise ValueError(f"No parameter named {name}")
-
-    def to_c(
-        self, degree: Optional[int] = None, dist_type: Optional[DistType] = None
-    ) -> np.ndarray:
-
-        priors = []
-        for prior in self.priors:
-            priors.append(prior.numeric_list())
-
-        # remove degreeN; 1st order multistage/polynomial
-        if degree and degree == 1:
-            priors.pop(2)
-
-        # copy degreeN; > 2rd order poly
-        if degree and degree > 2:
-            for i in range(2, degree):
-                priors.append(priors[2])
-
-        # add constant variance parameter
-        if dist_type and dist_type in {DistType.normal, DistType.log_normal}:
-            priors.append(self.variance_priors[0].numeric_list())
-
-        # add non-constant variance parameter
-        if dist_type and dist_type is DistType.normal_ncv:
-            for prior in self.variance_priors:
-                priors.append(prior.numeric_list())
-
-        return np.array(priors, dtype=np.float64).flatten("F")
